test(compiler): scope stored-callable process boundaries
